+++ src/Model/Produkt.java
@@ -3,7 +3,7 @@ package Model;
 public class Produkt {
 
     private String name;
-    private int price;
+    private double price;
     private String region;
 
     @Override
@@ -23,11 +23,11 @@ public class Produkt {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
@@ -39,7 +39,7 @@ public class Produkt {
         this.region = region;
     }
 
-    public Produkt(String name, int price, String region) {
+    public Produkt(String name, double price, String region) {
         this.name = name;
         this.price = price;
         this.region = region;
